+++ ACadSharp/IO/DWG/DwgStreamWriters/DwgObjectWriter.Objects.cs
@@ -1,0 +1,78 @@
+﻿using ACadSharp.Objects;
+using System;
+using System.Linq;
+
+namespace ACadSharp.IO.DWG
+{
+	internal partial class DwgObjectWriter : DwgSectionIO
+	{
+		private void writeObjects()
+		{
+			while (this._objects.Any())
+			{
+				CadObject obj = this._objects.Dequeue();
+
+				this.writeObject(obj);
+			}
+		}
+
+		private void writeObject(CadObject obj)
+		{
+			switch (obj)
+			{
+				case CadDictionary dictionary:
+					this.writeDictionary(dictionary);
+					break;
+				default:
+					throw new NotImplementedException($"Object not implemented : {obj.GetType().FullName}");
+			}
+		}
+
+		private void writeDictionary(CadDictionary dictionary)
+		{
+			this.writeCommonNonEntityData(dictionary);
+
+			//Common:
+			//Numitems L number of dictonary items
+			this._writer.WriteBitLong(dictionary.Count());
+
+			//R14 Only:
+			if (this._version == ACadVersion.AC1014)
+			{
+				//Unknown R14 RC Unknown R14 byte, has always been 0
+				this._writer.WriteByte(0);
+			}
+
+			//R2000 +:
+			if (this.R2000Plus)
+			{
+				//Cloning flag BS 281
+				this._writer.WriteBitShort((short)dictionary.ClonningFlags);
+				this._writer.WriteByte((byte)(dictionary.HardOwnerFlag ? 1u : 0u));
+			}
+
+			//Common:
+			foreach (var name in dictionary.EntryNames)
+			{
+				this._writer.WriteVariableText(name);
+			}
+
+			foreach (var handle in dictionary.EntryHandles)
+			{
+				this._writer.HandleReference(DwgReferenceType.SoftOwnership, handle);
+			}
+
+			this.registerObject(dictionary);
+
+			this.addEntriesToWriter(dictionary);
+		}
+
+		private void addEntriesToWriter(CadDictionary dictionary)
+		{
+			foreach (CadObject e in dictionary)
+			{
+				this._objects.Enqueue(e);
+			}
+		}
+	}
+}
